+++ how-to/integrate-with-salesforce/client/src/salesforce-integration.ts
@@ -270,7 +270,8 @@ export class SalesforceIntegration {
 		if (result.action.trigger === "focus-change") {
 			const data = result.data as SalesforceResultData;
 			if (data.obj && this._integrationHelpers) {
-				const palette = await this._integrationHelpers.getCurrentPalette();
+				const themeClient = await this._integrationHelpers.getThemeClient();
+				const palette = await themeClient.getPalette();
 				const resultWithReferences = await this.buildSearchResult(
 					this._integrationHelpers.templateHelpers,
 					palette,