+++ how-to/workspace-platform-starter/public/common/popups/menu/menu.js
@@ -22,32 +22,54 @@ async function initDOM() {
 	if (menuContainer) {
 		if (Array.isArray(options.customData?.menuEntries) && options.customData.menuEntries.length > 0) {
 			for (const menuEntry of options.customData.menuEntries) {
-				const menuEntryButton = document.createElement('div');
-				menuEntryButton.classList.add('menu-item');
-				menuEntryButton.addEventListener('click', async () => {
-					await fin.me.dispatchPopupResult(menuEntry.customData);
-				});
-				menuContainer.append(menuEntryButton);
+				if (menuEntry.type === 'separator') {
+					const menuEntryItem = document.createElement('div');
 
-				let icon;
-				if (menuEntry.icon) {
-					icon = document.createElement('img');
-					icon.src = menuEntry.icon;
-					icon.width = 16;
-					icon.height = 16;
+					menuEntryItem.classList.add('menu-item-separator');
+					if (!(menuEntry.enabled ?? true)) {
+						menuEntryItem.classList.add('menu-item-disabled');
+					}
+
+					const sep = document.createElement('hr');
+					menuEntryItem.append(sep);
+
+					menuContainer.append(menuEntryItem);
+				} else {
+					const menuEntryItem = document.createElement('button');
+					menuEntryItem.ariaLabel = menuEntry.label;
+
+					menuEntryItem.classList.add('menu-item');
+					if (!(menuEntry.enabled ?? true)) {
+						menuEntryItem.classList.add('menu-item-disabled');
+					} else {
+						menuEntryItem.addEventListener('click', async () => {
+							await fin.me.dispatchPopupResult(menuEntry.customData);
+						});
+					}
+
+					let icon;
+					if (menuEntry.icon) {
+						icon = document.createElement('img');
+						icon.src = menuEntry.icon;
+						icon.width = 16;
+						icon.height = 16;
+					}
+
+					const text = document.createElement('span');
+					text.textContent = menuEntry.label;
+
+					if (icon) {
+						menuEntryItem.append(icon);
+					}
+					menuEntryItem.append(text);
+					menuContainer.append(menuEntryItem);
 				}
-
-				const text = document.createElement('span');
-				text.textContent = menuEntry.label;
-
-				if (icon) {
-					menuEntryButton.append(icon);
-				}
-				menuEntryButton.append(text);
 			}
 		} else {
 			const noEntries = document.createElement('div');
-			noEntries.textContent = options.customData?.noEntryText ?? 'No entries';
+			const content = options.customData?.noEntryText ?? 'No entries';
+			noEntries.textContent = content;
+			noEntries.ariaLabel = content;
 			noEntries.classList.add('menu-item-none');
 			menuContainer.append(noEntries);
 		}